requirements.txt
Flask==2.0.1
pyTelegramBotAPI==3.8.1
numpy==1.21.1
Pillow==8.3.1